move EF entities into a entity namespace
