+++ services/Skyra.Database/Models/Entities/TwitchStreamSubscription.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 #nullable disable
 
-namespace Skyra.Database.Models.Generated
+namespace Skyra.Database.Models.Entities
 {
 	[Table("twitch_stream_subscription")]
 	public class TwitchStreamSubscription
