+++ services/Skyra.Database/Models/Entities/UserCooldown.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 #nullable disable
 
-namespace Skyra.Database.Models.Generated
+namespace Skyra.Database.Models.Entities
 {
 	[Table("user_cooldown")]
 	public class UserCooldown
