+++ services/Skyra.Database/Models/Entities/UserGameIntegration.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 #nullable disable
 
-namespace Skyra.Database.Models.Generated
+namespace Skyra.Database.Models.Entities
 {
 	[Table("user_game_integration")]
 	public class UserGameIntegration
